Wut, float != CGFloat

+++ SpringBoard/SBIconAnimator.h
@@ -11,7 +11,7 @@
 
 @interface SBIconAnimator : NSObject
 {
-    float _fraction;
+    CGFloat _fraction;
     BOOL _startedAnimation;
     BOOL _startAnimationAfterRotationEnds;
     BOOL _windowIsRotating;
@@ -34,17 +34,17 @@
 - (void)_windowFinishedRotating;
 - (void)_invalidateCompletions;
 - (id)_centralAnimationFactory;
-- (void)_animateToFraction:(float)arg1 afterDelay:(double)arg2 withSharedCompletion:(id)arg3;
+- (void)_animateToFraction:(CGFloat)arg1 afterDelay:(double)arg2 withSharedCompletion:(id)arg3;
 - (unsigned int)_numberOfSignificantAnimations;
 - (void)_cleanupAnimation;
-- (void)_setAnimationFraction:(float)arg1;
+- (void)_setAnimationFraction:(CGFloat)arg1;
 - (void)_prepareAnimation;
 - (BOOL)_isDelayedForRotation;
 - (void)_animateToFractionFromContext:(id)arg1;
 - (void)_animateToFractionFromPendingContexts;
 - (void)cleanup;
-- (void)animateToFraction:(float)arg1 afterDelay:(double)arg2 withCompletion:(void(^)(void))arg3;
-- (void)setFraction:(float)arg1;
+- (void)animateToFraction:(CGFloat)arg1 afterDelay:(double)arg2 withCompletion:(void(^)(void))arg3;
+- (void)setFraction:(CGFloat)arg1;
 - (void)prepare;
 - (id)centralAnimationFactory;
